Deleted stocks app and finshed index with pagination
realized I was creating a ton of overhead by creating a seperate stocks app only to add and index view of the stocks that are preloaded, so I deleted the app since I was only using one view.

CSS still needs to be added to match the websites theme.
