--- sarcastic-sparrows/sparrow_django/stocks/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class StocksConfig(AppConfig):
-    name = 'sparrow_django.stocks'
