--- sarcastic-sparrows/sparrow_django/stocks/models.py
@@ -1,3 +1,0 @@
-from django.db import models
-
-
