--- sarcastic-sparrows/sparrow_django/stocks/templates/stocks/index.html
@@ -1,27 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Index</title>
-</head>
-<body>
-    <table class="table table-bordered">
-        <thead>
-            <tr>
-                <th>Stock</th>
-                <th>Ticker</th>
-                <th>Date</th>
-            </tr>
-        </thead>
-        <tbody>
-            {% for stocks in Stock %}
-                <tr>
-                    <td>{{ stocks.id }}</td>
-                    <td>{{ stocks.ticker_symbol }}</td>
-                    <td>{{ stocks.date }}</td>
-                </tr>
-            {% endfor %}
-        </tbody>
-</table>
-</body>
-</html>
