--- sarcastic-sparrows/sparrow_django/stocks/urls.py
@@ -1,9 +1,0 @@
-from django.urls import path
-
-from . import views
-
-
-app_name = 'stocks'
-urlpatterns = [
-    path(" ", views.IndexView, name='index'),
-]
